+++ frontend/antlr4/Antlr4CSTVisitor.cpp
@@ -961,7 +961,7 @@ std::any MiniCCSTVisitor::visitNumber(MiniCParser::NumberContext * ctx)
     }
     else if (ctx->T_FLOAT_DIGIT()) {
     	   //TODO: [交流:fp]了解sysY的float标准
-        float_t val = (float_t) stoull(ctx->T_FLOAT_DIGIT()->getText());
+        float_t val = (float_t) stof(ctx->T_FLOAT_DIGIT()->getText());
         int64_t lineNo = (int64_t) ctx->T_FLOAT_DIGIT()->getSymbol()->getLine();
         //TODO: [fp]ast_node::New 扩展digit_float_attr
         numberNode = ast_node::New(digit_float_attr{val, lineNo});
